Ajout désactiver des utilisateurs

+++ src/Controller/UserController.php
@@ -2,14 +2,14 @@
 
 namespace App\Controller;
 
-use App\Entity\Sortie;
+use App\Entity\EtatSortie;
 use App\Entity\User;
+use App\Form\AnnulerSortieType;
 use App\Form\UserType;
+use App\Repository\SortieRepository;
 use App\Repository\UserRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\Form\Extension\Core\Type\FileType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -100,6 +100,12 @@ class   UserController extends AbstractController
      */
     public function supprimer(int $id, EntityManagerInterface $entityManager, Request $request)
     {
+        $userCurrent = $entityManager->getRepository(User::class)->find($id);
+        if($userCurrent->getAdministrateur()){
+            $this->addFlash('error', 'Vous ne pouvez pas supprimer un administrateur');
+            return $this->redirectToRoute('user_supprimer_utilisateurs');
+        }
+
         $user = $entityManager->getRepository(User::class)->find($id);
 
         $entityManager->remove($user);
@@ -125,7 +131,30 @@ class   UserController extends AbstractController
         return $this->render("/user/supprimerUtilisateurs.html.twig",[
            "allUsers" => $allUsers
         ]);
+    }
 
+    /**
+     * @Route("/desactiver/{id}", name="desactiver_utilisateur")
+     */
+    public function desactiver(int $id, EntityManagerInterface $entityManager, UserRepository $userRepository, SortieRepository $sortieRepository)
+    {
+        $user = $entityManager->getRepository(User::class)->find($id);
+
+        $userSorties = $sortieRepository->findBy(array('participantOrganisateur' => $id));
+        $sortieAnnulee = $entityManager->getRepository(EtatSortie::class)->find(4);
+
+        foreach($userSorties as $sortie){
+            $description = $sortie->getDescription();
+            $sortie->setDescription("SORTIE ANNULEE - ".$description);
+            $sortie->setEtatSortie($sortieAnnulee);
+           // $user->setRoles(["ROLE_DESACTIVE"]);
+        }
+
+        $user->setActif(false);
+        $entityManager->persist($user);
+        $entityManager->flush();
+
+        return $this->redirectToRoute('user_supprimer_utilisateurs');
 
     }
